Loki: Move `pragma_utils` into `loki.ir`
