+++ loki/backend/dacegen.py
@@ -7,7 +7,7 @@
 
 from loki.backend.pygen import PyCodegen
 from loki.expression import symbols as sym, ExpressionRetriever
-from loki.pragma_utils import is_loki_pragma
+from loki.ir import is_loki_pragma
 from loki.types import BasicType
 
 __all__ = ['dacegen', 'DaceCodegen']
