+++ loki/transform/transform_inline.py
@@ -18,13 +18,12 @@ from loki.expression import (
 )
 from loki.ir import (
     Import, Comment, Assignment, VariableDeclaration, CallStatement,
-    Transformer, FindNodes
+    Transformer, FindNodes, pragmas_attached, is_loki_pragma
 )
 from loki.expression import symbols as sym
 from loki.types import BasicType
 from loki.tools import as_tuple
 from loki.logging import warning, error
-from loki.pragma_utils import pragmas_attached, is_loki_pragma
 from loki.subroutine import Subroutine
 
 from loki.transform.transform_sanitise import transform_sequence_association_append_map
