+++ loki/transform/transform_region.py
@@ -14,11 +14,12 @@ from collections import defaultdict
 from loki.analyse import dataflow_analysis_attached
 from loki.expression import FindVariables, Variable
 from loki.ir import (
-    CallStatement, Comment, Import, Loop, Pragma, PragmaRegion, Section,
-    FindNodes, FindScopes, MaskedTransformer, NestedMaskedTransformer, Transformer
+    CallStatement, Comment, Import, Loop, Pragma, PragmaRegion,
+    Section, FindNodes, FindScopes, MaskedTransformer,
+    NestedMaskedTransformer, Transformer, is_loki_pragma,
+    get_pragma_parameters, pragma_regions_attached
 )
 from loki.logging import info
-from loki.pragma_utils import is_loki_pragma, get_pragma_parameters, pragma_regions_attached
 from loki.subroutine import Subroutine
 from loki.tools import as_tuple, flatten, CaseInsensitiveDict
 from loki.transform.transform_array_indexing import (
